docs: extract header and footer from articleDemoPage
